refactor(vector/index): extract shared multi-segment fan-out search layer (#902)
HNSW's segmented search fan-out (SegmentedReaderFacade, the newest-wins
containment masking + #883 expanding-refill searcher) is implemented
entirely against VectorIndexReader/VectorIndexSearcher trait methods —
it never depended on any HNSW-specific type. The one near-miss was
shadowed()'s r.vectors().contains(doc_id, field), a direct call into
HNSW's internal quantized-vector pool; it is semantically identical to
the trait method contains_vector(doc_id, field), which Flat's and
IVF's readers already implement the same way.

Moves SegmentedReaderFacade (generalized to Vec<Arc<dyn
VectorIndexReader>>) and the fan-out searcher (renamed
SegmentFanoutSearcher, taking per-segment concrete-searcher
construction as a closure instead of hard-coding HnswSearcher) into
segment/fanout.rs. HNSW's reader()/searcher() now coerce their
concrete Vec<Arc<HnswIndexReader>> into the trait-object form and
supply a closure capturing ef_search.

Zero logic changes to the masking/expanding-refill/count algorithms
themselves. Verified via the existing regression net: both
vector_segmented_index_test.rs (16 tests) and
vector_segmented_correctness_test.rs (5 tests) — confirmed
byte-for-byte unedited via `git diff --stat` before running — pass in
full, including multi_segment_self_recall_matches_monolithic and
adaptive_refill_recovers_live_doc_behind_stale_band, the two most
load-bearing pins in this campaign.

This was the highest-risk refactor in the #889 campaign (retypes a hot
search path), kept as its own PR for focused review.

Part of #889
Closes #902
